simplify app-web footprint in app-api
