+++ services/app-api/src/common-code/healthPlanFormDataType/healthPlanFormData.ts
@@ -1,4 +1,4 @@
-import {
+import type {
     RateInfoType,
     UnlockedHealthPlanFormDataType,
     ActuaryContact,
@@ -7,7 +7,7 @@ import {
     provisionCHIPKeys,
     modifiedProvisionMedicaidAmendmentKeys,
     modifiedProvisionMedicaidBaseKeys,
-    GeneralizedModifiedProvisions,
+    type GeneralizedModifiedProvisions,
 } from './ModifiedProvisions'
 import { generateApplicableProvisionsList } from '../healthPlanSubmissionHelpers/provisions'
 import { formatRateNameDate } from '../../common-code/dateHelpers'
@@ -15,7 +15,7 @@ import type { LockedHealthPlanFormDataType } from './LockedHealthPlanFormDataTyp
 import type { HealthPlanFormDataType } from './HealthPlanFormDataType'
 import type { ProgramArgType } from '.'
 import {
-    CHIPFederalAuthority,
+    type CHIPFederalAuthority,
     federalAuthorityKeysForCHIP,
 } from './FederalAuthorities'
 
@@ -96,32 +96,32 @@ const hasValidActuaries = (actuaries: ActuaryContact[]): boolean =>
 
 const hasValidRates = (sub: LockedHealthPlanFormDataType): boolean => {
     const validRates = sub.rateInfos.every((rateInfo) => {
-                  const validBaseRate =
-                      rateInfo.rateType !== undefined &&
-                      rateInfo.rateDateCertified !== undefined &&
-                      rateInfo.rateDateStart !== undefined &&
-                      rateInfo.rateDateEnd !== undefined &&
-                      rateInfo.rateProgramIDs !== undefined &&
-                      rateInfo.rateProgramIDs.length > 0 &&
-                      hasValidActuaries(rateInfo.actuaryContacts)
+        const validBaseRate =
+            rateInfo.rateType !== undefined &&
+            rateInfo.rateDateCertified !== undefined &&
+            rateInfo.rateDateStart !== undefined &&
+            rateInfo.rateDateEnd !== undefined &&
+            rateInfo.rateProgramIDs !== undefined &&
+            rateInfo.rateProgramIDs.length > 0 &&
+            hasValidActuaries(rateInfo.actuaryContacts)
 
-                  if (isRateAmendment(rateInfo)) {
-                      return (
-                          validBaseRate &&
-                          Boolean(
-                              rateInfo.rateAmendmentInfo &&
-                                  rateInfo.rateAmendmentInfo.effectiveDateEnd &&
-                                  rateInfo.rateAmendmentInfo.effectiveDateStart
-                          )
-                      )
-                  }
+        if (isRateAmendment(rateInfo)) {
+            return (
+                validBaseRate &&
+                Boolean(
+                    rateInfo.rateAmendmentInfo &&
+                        rateInfo.rateAmendmentInfo.effectiveDateEnd &&
+                        rateInfo.rateAmendmentInfo.effectiveDateStart
+                )
+            )
+        }
 
-                  return validBaseRate
-              })
+        return validBaseRate
+    })
 
     // Contract only - skip all validations for hasValidRates
     if (sub.submissionType === 'CONTRACT_ONLY') {
-       return true
+        return true
     } else {
         return validRates
     }
