--- services/app-api/src/constants/index.ts
@@ -1,31 +1,0 @@
-export {
-    ModifiedProvisionsCHIPRecord,
-    ModifiedProvisionsBaseContractRecord,
-    ModifiedProvisionsAmendmentRecord,
-} from './modifiedProvisions'
-
-export {
-    ContractTypeRecord,
-    SubmissionTypeRecord,
-    FederalAuthorityRecord,
-    ManagedCareEntityRecord,
-    ActuaryFirmsRecord,
-    ActuaryCommunicationRecord,
-    ContractExecutionStatusRecord,
-    SubmissionStatusRecord,
-    PopulationCoveredRecord,
-} from './healthPlanPackages'
-
-export {
-    PageHeadingsRecord,
-    PageTitlesRecord,
-    RoutesRecord,
-    ROUTES,
-    STATE_SUBMISSION_FORM_ROUTES,
-    STATE_SUBMISSION_SUMMARY_ROUTES,
-    QUESTION_RESPONSE_SHOW_SIDEBAR_ROUTES,
-} from './routes'
-
-export type { RouteT, RouteTWithUnknown } from './routes'
-
-export { ERROR_MESSAGES } from './errors'
