--- services/app-api/src/constants/statutoryRegulatoryAttestation.ts
@@ -1,16 +1,0 @@
-const StatutoryRegulatoryAttestation = {
-    YES: 'Yes, the contract fully complies with all applicable requirements',
-    NO: 'No, the contract does not fully comply with all applicable requirements',
-}
-
-const StatutoryRegulatoryAttestationQuestion =
-    'Do you attest that this contract complies with all applicable statutory and regulatory requirements including those contained in Title 42 Part 438 and Part 457 of the Code of Federal Regulations (CFR)?'
-
-const StatutoryRegulatoryAttestationDescription =
-    'Provide a brief description of any contractual or operational non-compliance, including regulatory citations and expected timeframe for remediation'
-
-export {
-    StatutoryRegulatoryAttestation,
-    StatutoryRegulatoryAttestationQuestion,
-    StatutoryRegulatoryAttestationDescription,
-}
