--- services/app-api/src/documentHelpers/index.ts
@@ -1,2 +1,0 @@
-export { makeDocumentDateTable } from './makeDocumentDateLookupTable'
-export { makeDocumentS3KeyLookup } from './makeDocumentKeyLookupList'
