--- services/app-api/src/otelHelpers/index.ts
@@ -1,7 +1,0 @@
-export {
-    getTracer,
-    recordSpan,
-    recordJSException,
-    recordJSExceptionWithContext,
-    recordUserInputException,
-} from './tracingHelper'
